requirements.txt
alembic==1.4.2
certifi==2018.11.29
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
requests==2.23.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
Werkzeug==1.0.1
